LayoutTests/fast/files/url-required-arguments.html

<!DOCTYPE HTML>
<html>
 <head>
    <script src="../js/resources/js-test-pre.js"></script>
 </head>
 <body>
    <div id="description"></div>
    <div id="console"></div>
<script>
description("Test required arguments of URL object.");

shouldThrow("URL.createObjectURL()");
shouldThrow("URL.revokeObjectURL()");

</script>
    <script src="../js/resources/js-test-post.js"></script>
 </body>
</html>
